fixed all the bugs and also the float values;

+++ core/data_formats/typed_dicts/distributor_typdict.py
@@ -4,11 +4,11 @@ from datetime import date
 
 class DistributorDiscountsTypDict(TypedDict):
     discount:str
-    minimum_thershold:int
+    minimum_thershold:float
     rebate_type:str
 
 class DistributorPaymentInfosTypDict(TypedDict):
     invoice_number:Optional[Union[None,str]]=None
-    invoice_value:Optional[Union[int,None]]=None
-    paid_amount:Optional[Union[None,int]]=None
+    invoice_value:Optional[Union[float,None]]=None
+    paid_amount:Optional[Union[None,float]]=None
     paid_date:Optional[Union[None,date]]=None
