+++ core/data_formats/typed_dicts/order_typdict.py
@@ -15,7 +15,7 @@ class StatusInfo(TypedDict):
     invoice_status:Union[str,InvoiceStatus]
     invoice_number:NotRequired[Union[str,None]]=None
     invoice_date:NotRequired[Union[date,None]]=None
-    paid_amount:NotRequired[Union[int,None]]=None
+    paid_amount:NotRequired[Union[float,None]]=None
 
 class LogisticsInfo(TypedDict):
     purchase_type:Union[str,PurchaseTypes]
